+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `AttributeName::from_str_unchecked`.
+
 ### Removed
 
 - Removed `AttributeName::new`.